pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "antagents"
version = "0.1.dev"
description = "🐜 AntAgents: Lightweight Library for Agent Swarms"
authors = [
  { name="JW", email="thomas.junjie.wang@qq.com" },
]
readme = "README.md"
requires-python = ">=3.11"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["antagents"]

[project.scripts]
antagent = "antagents.cli:main"

[tool.poetry.dependencies]
ipython = "9.0.2"
jinja2 = "3.1.6"
markdownify = "1.1.0"
mcp = "1.13.1"
mcpadapt = "0.1.13"
numpy = "2.3.2"
openai = "1.102.0"
pillow = "11.3.0"
protobuf = "6.32.0"
python-dotenv = "1.1.1"
pyyaml = "6.0.2"
requests = "2.32.5"
rich = "14.1.0"
soundfile = "0.13.1"
wikipedia-api = "0.8.1"
